fix sending sms by meetingId

+++ public/app/components/candidates/components/candidate-meetings/candidate-meetings.client.comp.es6
@@ -15,12 +15,13 @@ angular
     });
 
 /* @ngInject */
-function candidateMeetingsCtrl(NgTableParams, $state, candidateMeetingsService, interviewersService, toastr) {
+function candidateMeetingsCtrl(NgTableParams, $state, candidateMeetingsService, interviewersService, toastr, yesNoModalService, smsService) {
     let $ctrl = this;
     // $ctrl.addMeeting = addMeeting;
     $ctrl.editMeeting = editMeeting;
     $ctrl.getInterviewers = getInterviewers;
     $ctrl.deleteMeeting = deleteMeeting;
+    $ctrl.sendSMS = sendSMS;
 
     $ctrl.$onInit = () => {
         candidateMeetingsService.getMeetings($ctrl.candidate.candidateId)
@@ -49,10 +50,6 @@ function candidateMeetingsCtrl(NgTableParams, $state, candidateMeetingsService, 
 
         return new NgTableParams(initialParams, initialSettings);
     }
-    //
-    // function addMeeting() {
-    //     $state.go("meetingEditor", {candidateId: $ctrl.candidate.candidateId});
-    // }
 
     function editMeeting(meetingId) {
         $state.go("meetingEditor", {candidateId: $ctrl.candidate.candidateId, meetingId: meetingId});
@@ -63,16 +60,37 @@ function candidateMeetingsCtrl(NgTableParams, $state, candidateMeetingsService, 
     }
 
     function deleteMeeting(meetingId) {
-        candidateMeetingsService.deleteMeeting(meetingId, $ctrl.candidate.candidateId)
-            .then(() => {
-                toastr.success('Meeting has been deleted', "Delete");
-                $ctrl.meetingsTableParams.settings().dataset = _.reject($ctrl.meetingsTableParams.settings().dataset, {meetingId: meetingId});
-                $ctrl.meetingsTableParams.reload();
-                candidateMeetingsService.removeCandidateMeetingCache($ctrl.candidate.candidateId);
-            })
-            .catch(response => {
-                toastr.error("Cannot delete meeting", "Delete Error");
-                console.error("Cannot delete meeting, err: " + response.data.message);
+        let modalInstance = yesNoModalService.createModal('Confirm Delete', "Are you sure you want to delete this meeting?");
+
+        modalInstance.result
+            .then(approve => {
+                if (approve) {
+                    candidateMeetingsService.deleteMeeting(meetingId, $ctrl.candidate.candidateId)
+                        .then(() => {
+                            toastr.success('Meeting has been deleted', "Delete");
+                            $ctrl.meetingsTableParams.settings().dataset = _.reject($ctrl.meetingsTableParams.settings().dataset, {meetingId: meetingId});
+                            $ctrl.meetingsTableParams.reload();
+                            candidateMeetingsService.removeCandidateMeetingCache($ctrl.candidate.candidateId);
+                        })
+                        .catch(response => {
+                            toastr.error("Cannot delete meeting", "Delete Error");
+                            console.error("Cannot delete meeting, err: " + response.data.message);
+                        });
+                }
+            });
+    }
+
+    function sendSMS(meetingId) {
+        let modalSms = yesNoModalService.createModal('Confirm SMS', `Are you sure you want to send SMS ${$ctrl.candidate.firstName} ${$ctrl.candidate.lastName}?`);
+
+        modalSms.result
+            .then(approve => {
+                if (approve) {
+                    smsService.sendSMS(meetingId)
+                        .then(() => {
+                            toastr.success(`SMS was sent to candidate ${$ctrl.candidate.firstName} ${$ctrl.candidate.lastName}`, "SMS");
+                        });
+                }
             });
     }
 }
